+++ src/utils/logger.ts
@@ -1,4 +1,4 @@
-import chalk from 'chalk'
+import * as chalk from 'chalk'
 
 const LOG_TYPES = {
   NONE: 0,
@@ -17,17 +17,17 @@ const logTime = () => {
 
 const log = (...args) => {
   if (logType < LOG_TYPES.NORMAL) return
-  console.log(logTime(), process.pid, chalk.bold.green('[INFO]'), ...args)
+  console.log(logTime(), process.pid, chalk.greenBright.bold('[INFO]'), ...args)
 }
 
 const error = (...args) => {
   if (logType < LOG_TYPES.ERROR) return
-  console.log(logTime(), process.pid, chalk.bold.red('[ERROR]'), ...args)
+  console.log(logTime(), process.pid, chalk.redBright.bold('[ERROR]'), ...args)
 }
 
 const debug = (...args) => {
   if (logType < LOG_TYPES.DEBUG) return
-  console.log(logTime(), process.pid, chalk.bold.blue('[DEBUG]'), ...args)
+  console.log(logTime(), process.pid, chalk.blueBright.bold('[DEBUG]'), ...args)
 }
 
-export default { LOG_TYPES, log, error, debug }
+export default { log, error, debug }