+++ src/utils/mongoose.ts
@@ -1,6 +1,7 @@
 import * as mongoose from 'mongoose'
 
-import { highlight } from './chalk'
+import * as hl from 'utils/chalk'
+import Logger from 'utils/logger'
 
 export async function connectMongoDB(mongoUrl: string) {
   mongoose.set('useCreateIndex', true)
@@ -12,10 +13,10 @@ export async function connectMongoDB(mongoUrl: string) {
       useUnifiedTopology: true,
     })
     .then(() => {
-      console.log(`${highlight(0, '✓')} Database connected (URL: ${highlight(0, mongoUrl)})`)
+      Logger.log(`Database connected (URL: ${hl.success(mongoUrl)})`)
     })
     .catch((err) => {
-      console.log(highlight(3, `Database connection: ${err.message}`))
+      Logger.error(hl.error(`Database connection: ${err.message}`))
       process.exit(0)
     })
 }